requirements.txt
discord.py
python-dotenv
aiocron
pytz
